spark/requirements.txt
requests>=2.31.0
pyyaml>=6.0
rich>=13.0.0
